refactor: maxAvailableHomeCountFをHome domainへ移動
